+++ psana/psana/pyalgos/generic/Entropy.py
@@ -1,6 +1,6 @@
 #!@PYTHON@
 ####!/usr/bin/env python
-#------------------------------
+
 """
 :py:class:`Entropy.py` - collection of methods to evaluate data array entropy
 =============================================================================
@@ -15,8 +15,8 @@ Usage::
     from pyalgos.generic.NDArrGenerators import random_standard
     from pyalgos.generic.NDArrUtils import print_ndarr
 
-    arr_float = random_standard(shape=(1000), mu=200, sigma=25, dtype=np.float)
-    arr_int16 = arr_float.astype(np.int16)  
+    arr_float = random_standard(shape=(1000), mu=200, sigma=25, dtype=np.float32)
+    arr_int16 = arr_float.astype(np.int16)
     print_ndarr(arr_int16, name='arr_int16', first=0, last=10)
 
     print 'entropy(arr_int16)     = %.6f' % entropy(arr_int16)
@@ -33,9 +33,8 @@ If you use all or part of it, please give an appropriate acknowledgment.
 
 Created by Mikhail Dubrovin
 """
-#------------------------------
+
 import numpy as np
-#------------------------------
 
 def hist_values(nda) :
     """Depending on nda.dtype fills/returns 1-D 2^8(16)-bin histogram-array of 8(16)-bit values of input n-d array
@@ -49,13 +48,12 @@ def hist_values(nda) :
         return np.bincount(nda.flatten(), weights=None, minlength=1<<16)
 
     elif nda.dtype == np.int16 :
-        unda = nda.astype(np.uint16) # int16 (1,2,-3,0,4,-5,...) -> uint16 (1,2,0,65533,4,65531,...)        
+        unda = nda.astype(np.uint16) # int16 (1,2,-3,0,4,-5,...) -> uint16 (1,2,0,65533,4,65531,...)
         return np.bincount(unda.flatten(), weights=None, minlength=1<<16)
 
-    else : 
+    else :
         sys.exit('method %s get unexpected nda dtype=%s. Use np.uint8 or np.(u)int16'%(FR().f_code.co_name, str(nda.dtype)))
 
-#------------------------------
 
 def hist_probabilities(nda) :
     """Returns histogram-array of probabilities for each of (u)int16, uint8 intensity
@@ -68,15 +66,14 @@ def hist_probabilities(nda) :
     #print('Check sum of probabilities: %.6f for number of values in array = %d' % (ph.sum(), nvals))
     return ph
 
-#------------------------------
 
 def entropy(nda) :
     """Evaluates n-d array entropy using formula from https://en.wikipedia.org/wiki/Entropy_%28information_theory%29
-    """ 
+    """
     unda = None # histogram array indexes must be unsigned
     if   nda.dtype == np.uint8 : unda = nda
     elif nda.dtype == np.uint16: unda = nda
-    elif nda.dtype == np.int16 : unda = nda.astype(np.uint16) # int16 (1,2,-3,0,4,-5,...) -> uint16 (1,2,0,65533,4,65531,...)        
+    elif nda.dtype == np.int16 : unda = nda.astype(np.uint16) # int16 (1,2,-3,0,4,-5,...) -> uint16 (1,2,0,65533,4,65531,...)
 
     prob_h = hist_probabilities(unda)
 
@@ -89,34 +86,31 @@ def entropy(nda) :
     #print_ndarr(p_log2p_nda, name='per pixel array of P*log2(P)\n', first=1000, last=1010)
     return ent
 
-#------------------------------
 ## formula in https://en.wikipedia.org/wiki/Entropy_%28information_theory%29
 ## sums over all (x_i) which is a set of possible values....
-## this method sums over set (one entry) of probabilities 
-#------------------------------
+## this method sums over set (one entry) of probabilities
 
 def entropy_v1(nda) :
     """The same as entropy(nda) in a single place.
     """
     #print('%s for array dtype=%s'%(FR().f_code.co_name, str(nda.dtype)))
 
-    unda = nda 
+    unda = nda
     if   nda.dtype == np.uint8  : unda = nda
     elif nda.dtype == np.uint16 : unda = nda
-    elif nda.dtype == np.int16  : unda = nda.astype(np.uint16) # int16 (1,2,-3,0,4,-5,...) -> uint16 (1,2,0,65533,4,65531,...)        
+    elif nda.dtype == np.int16  : unda = nda.astype(np.uint16) # int16 (1,2,-3,0,4,-5,...) -> uint16 (1,2,0,65533,4,65531,...)
     else : sys.exit('method %s get unexpected nda dtype=%s. Use np.uint8 or np.(u)int16'%(FR().f_code.co_name, str(nda.dtype)))
 
     hsize = (1<<8) if nda.dtype == np.uint8 else (1<<16)
     vals_h = np.bincount(unda.flatten(), weights=None, minlength=hsize)
-    prob_h = np.array(vals_h, dtype=np.float) / unda.size
-    #prob_nda = prob_h[unda]    
+    prob_h = np.array(vals_h, dtype=np.float32) / unda.size
+    #prob_nda = prob_h[unda]
     #p_log2p_nda = prob_nda * np.log2(prob_nda)
     #ent = -p_log2p_nda.sum()
     p_log2p_nda = [p*np.log2(p) for p in prob_h if p>0]
     ent = -np.sum(p_log2p_nda)
     return ent
 
-#------------------------------
 
 def entropy_cpo(signal):
     '''Entropy evaluation method found by cpo on web
@@ -130,9 +124,6 @@ def entropy_cpo(signal):
     ent=np.sum([p*np.log2(1.0/p) for p in propab])
     return ent
 
-#------------------------------
-#------------------------------
-#------------------------------
 
 def test_entropy():
 
@@ -142,9 +133,9 @@ def test_entropy():
     from psana.pyalgos.generic.NDArrUtils import print_ndarr
     from time import time
 
-    arr_float = random_standard(shape=(100000,), mu=200, sigma=25, dtype=np.float)
+    arr_float = random_standard(shape=(100000,), mu=200, sigma=25, dtype=np.float32)
     arr_int16 = arr_float.astype(np.int16)
-    
+
     print_ndarr(arr_int16, name='arr_int16', first=0, last=10)
 
     t0_sec = time()
@@ -156,7 +147,6 @@ def test_entropy():
     print('entropy_v1(arr_int16)  = %.6f, time=%.6f sec' % (ent2, t2_sec-t1_sec))
     print('entropy_cpo(arr_int16) = %.6f, time=%.6f sec' % (ent3, t3_sec-t2_sec))
 
-#------------------------------
 
 def unitest_entropy():
     import sys
@@ -168,7 +158,6 @@ def unitest_entropy():
     print('entropy(arr_int16) = %.6f' % ent1)
     assert('%.6f'%ent1 == '6.690948')
 
-#------------------------------
 
 if __name__ == "__main__" :
     import sys; global sys
@@ -181,5 +170,5 @@ if __name__ == "__main__" :
     else : sys.exit('Test %s is not implemented' % tname)
     sys.exit('End of Test %s' % tname)
 
-#------------------------------
+# EOF
